Improve text editor placeholder and label logic
Updated the text editor to use an empty string as the default content instead of 'Enter Description' and added a placeholder prop. The label is now only rendered if showTitle and label are provided. Also, minor style and storybook control updates were made. Removed unused down arrow in time picker.